# Conflicts:
#	src/test/scala/io/qbeast/spark/index/model/transformer/TransformerIndexingTest.scala

+++ src/test/scala/io/qbeast/QbeastIntegrationTestSpec.scala
@@ -18,14 +18,9 @@ package io.qbeast
 import com.github.mrpowers.spark.fast.tests.DatasetComparer
 import io.qbeast.context.QbeastContext
 import io.qbeast.context.QbeastContextImpl
-import io.qbeast.core.keeper.Keeper
-import io.qbeast.core.keeper.LocalKeeper
 import io.qbeast.core.model.IndexManager
 import io.qbeast.core.model.QTableID
 import io.qbeast.core.model.QbeastSnapshot
-import io.qbeast.spark.delta.DeltaMetadataManager
-import io.qbeast.spark.delta.DeltaRollupDataWriter
-import io.qbeast.spark.delta.DeltaStagingDataManagerFactory
 import io.qbeast.spark.index.SparkColumnsToIndexSelector
 import io.qbeast.spark.index.SparkOTreeManager
 import io.qbeast.spark.index.SparkRevisionFactory
@@ -38,7 +33,9 @@ import org.apache.spark.SparkConf
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
+import java.io.File
 import java.nio.file.Files
+import scala.reflect.io.Directory
 
 /**
  * This class contains all function that you should use to test qbeast over spark. You can use it
@@ -89,6 +86,7 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
    * @param testCode
    *   the code to run within the spark session
    * @tparam T
+   *   the type of the output
    * @return
    */
   def withExtendedSpark[T](sparkConf: SparkConf = new SparkConf())(
@@ -116,6 +114,7 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
    * @param testCode
    *   the code to test within the spark session
    * @tparam T
+   *   the type of the output
    * @return
    */
   def withSpark[T](testCode: SparkSession => T): T = {
@@ -126,7 +125,9 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
    * Runs code with a Temporary Directory. After execution, the content of the directory is
    * deleted.
    * @param testCode
+   *   the test code
    * @tparam T
+   *   the test result type
    * @return
    */
   def withTmpDir[T](testCode: String => T): T = {
@@ -134,8 +135,14 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
     try {
       testCode(directory.toString)
     } finally {
-      import scala.reflect.io.Directory
-      val d = new Directory(directory.toFile)
+      removeDirectory(directory.toString)
+    }
+  }
+
+  def removeDirectory(directoryPath: String): Unit = {
+    val directory = new File(directoryPath)
+    if (directory.exists() && directory.isDirectory) {
+      val d = new Directory(directory)
       d.deleteRecursively()
     }
   }
@@ -144,11 +151,7 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
     withTmpDir(tmpDir => withSpark(spark => testCode(spark, tmpDir)))
 
   /**
-   * Runs a given test code with a QbeastContext instance. The specified Keeper instance is used
-   * to customize the QbeastContext.
-   *
-   * @param keeper
-   *   the keeper
+   * Runs a given test code with a QbeastContext instance.
    * @param testCode
    *   the test code
    * @tparam T
@@ -156,18 +159,15 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
    * @return
    *   the test result
    */
-  def withQbeastAndSparkContext[T](keeper: Keeper = LocalKeeper)(
-      testCode: SparkSession => T): T = {
+  def withQbeastAndSparkContext[T]()(testCode: SparkSession => T): T = {
     withSpark { spark =>
       val indexedTableFactory = new IndexedTableFactoryImpl(
-        keeper,
         SparkOTreeManager,
-        DeltaMetadataManager,
-        DeltaRollupDataWriter,
-        DeltaStagingDataManagerFactory,
+        QbeastContext.metadataManager,
+        QbeastContext.dataWriter,
         SparkRevisionFactory,
         SparkColumnsToIndexSelector)
-      val context = new QbeastContextImpl(spark.sparkContext.getConf, keeper, indexedTableFactory)
+      val context = new QbeastContextImpl(spark.sparkContext.getConf, indexedTableFactory)
       try {
         QbeastContext.setUnmanaged(context)
         testCode(spark)
@@ -180,7 +180,9 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
   /**
    * Runs code with QbeastContext and a Temporary Directory.
    * @param testCode
+   *   actual test code
    * @tparam T
+   *   output type
    * @return
    */
 
@@ -192,6 +194,7 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
    * @param testCode
    *   the code to reproduce
    * @tparam T
+   *   the type of the output
    * @return
    */
   def withQbeastContextSparkAndTmpWarehouse[T](testCode: (SparkSession, String) => T): T =
@@ -203,7 +206,9 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
   /**
    * Runs code with OTreeAlgorithm configuration
    * @param code
+   *   IndexManager
    * @tparam T
+   *   code type
    * @return
    */
   def withOTreeAlgorithm[T](code: IndexManager => T): T = {
